+++ cmd/docmgr/cmds/meta/update.go
@@ -3,6 +3,7 @@ package meta
 import (
 	"github.com/go-go-golems/docmgr/cmd/docmgr/cmds/common"
 	"github.com/go-go-golems/docmgr/pkg/commands"
+	"github.com/go-go-golems/glazed/pkg/cli"
 	"github.com/spf13/cobra"
 )
 
@@ -11,5 +12,9 @@ func newUpdateCommand() (*cobra.Command, error) {
 	if err != nil {
 		return nil, err
 	}
-	return common.BuildCommand(cmd)
+	return common.BuildCommand(
+		cmd,
+		cli.WithDualMode(true),
+		cli.WithGlazeToggleFlag("with-glaze-output"),
+	)
 }